Sort the user lists alphabetically
Sort the user lists alphabetically
Previously the user lists are semi random which leads to situations
like this:
-groupname = user1, user2, user3, user4
+groupname = user1, user3, user2, user4
As you can see nothing has actually changed here except the order in
which users are listed.
This is unacceptable if the produced user list is versioned in SVN and
then used directly from SVN using the following syntax in the
httpd.conf file
AuthzSVNGroupsFile
file:///path/to/svn/reponame/authz/svn.groups
This is also unacceptable in the case that the user list is versioned
in SVN and then used as an audit trail.