Add files via upload
Added a new option --user-query-template. If specified it will be used instead of the --user-query, and any {} placeholders in the string will be replaced by the results of the group member query. The motivation for this is using the script with our Open Directory, where a the group member query returns a string like "example.user" instead of a full DN "uid=example.user,ou=Users,dc=example,dc=com" - so it can't be directly used as a query.